Added security features to only allow superuser to edit/delete/add products

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, reverse, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -99,8 +100,13 @@ def collections(request):
     return render(request, template, context)
 
 
+@login_required
 def add_product(request):
     """ Add a new product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the website owner can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -120,8 +126,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit an existing product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the website owner can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -144,8 +155,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete an existing product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the website owner can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'{product.name} has been deleted!')
